requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
coverage==7.13.5
curl_cffi==0.13.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
ezdxf==1.4.3
fastapi==0.135.2
fonttools==4.62.1
fpdf2==2.8.7
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
kiwisolver==1.5.0
librt==0.8.1
lxml==6.0.2
macholib==1.16.4
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
multitasking==0.0.12
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.3
openai==2.29.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
peewee==4.0.2
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ruff==0.15.7
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
yfinance==1.2.0
